Moving tasks into separate files.
