+++ gulp/scss-lint.js
@@ -1,0 +1,14 @@
+var gulp        = require('gulp'),
+    paths       = require('../config').paths,
+    scsslint    = require('gulp-scss-lint');
+
+// Runs SCSS linter.
+// gulp link
+gulp.task('lint', function() {
+  gulp.src(paths.styles)
+    .pipe(scsslint({
+      'bundleExec': true,
+      'config': '.scss-lint.yml'
+    }))
+    .pipe(scsslint.failReporter());
+});
